requirements.txt
telethon>=1.36.0
aiohttp>=3.9.0
colorama>=0.4.6
python-dotenv>=1.0.0
Pillow>=10.0.0
aiosqlite>=0.19.0
python-socks>=2.4.0
cryptography>=41.0.0
psutil>=5.9.0
